requirements.txt
cycler
kiwisolver
matplotlib
numpy
Pillow
pyparsing
python-dateutil
pytz
six
torch
torchvision
